requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
Faker==12.0.1
importlib-metadata==4.11.3
Markdown==3.3.7
mixer==7.2.2
pyasn1==0.4.8
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
zipp==3.8.0
